requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.2.0
certifi==2024.2.2
contourpy==1.2.0
contrib==0.3.0
cycler==0.12.1
dnspython==2.6.1
fonttools==4.49.0
frozenlist==1.4.1
idna==3.6
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
load-dotenv==0.1.0
magic-filter==1.0.12
matplotlib==3.8.3
multidict==6.0.5
numpy==1.26.4
packaging==23.2
pillow==10.2.0
pydantic==2.5.3
pydantic_core==2.14.6
pymongo==4.6.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.33.0
rpds-py==0.18.0
six==1.16.0
typing_extensions==4.10.0
yarl==1.9.4
motor==3.3.2
aiogram-dialog==2.1.0
cachetools==5.3.3
jinja2==3.1.3